Add 3 new packages with 1 file each
Three new packages have been added, which contain one new file each.
There's an addition to "application.properties". The packages are:
"controllers", "entities" and "repositories". The files are:
"UsersController", "User" and "UsersRepository".

The goal of this addition is to see if everything is working as it
should be, meaning you can go to "localhost/users/listAll" and see if
there's a response.